Ajustes executados para domain e repository Customer
