+++ system_atm/domain/entities/customer.py
@@ -1,0 +1,7 @@
+class Customer:
+    
+    def __init__(self, nome: str, conta: str, saldo: float) -> None:
+
+        self.nome = nome
+        self.conta = conta
+        self.saldo = saldo
