+++ system_atm/infrastructure/repositories/customer_repository.py
@@ -1,0 +1,18 @@
+from domain.entities.customer import Customer
+
+def customer_respository(repository: dict) -> dict:
+
+    def create(customer: Customer) -> int:
+        query = """
+        INSERT INTO DADOS_BANCARIOS(NOME, CONTA, SALDO) \
+            VALUES(:NOME, :CONTA, :SALDO)
+        """
+        return repository["execute"](query, (customer.nome, customer.conta, customer.saldo)).rowcount
+
+    def find_by_id(id: int) -> tuple:
+
+        return repository["fetchone"]("""
+        SELECT * FROM DADOS_BANCARIOS WHERE ID=:ID
+        """, (id, ))
+
+    return {"create":create, "find_by_id":find_by_id}
