+++ system_atm/infrastructure/repositories/repository.py
@@ -1,0 +1,14 @@
+from config import Cursor
+
+def repository(cursor: Cursor) -> dict:
+
+    def execute(query: str, parameters: tuple):
+        return cursor.execute(query, parameters)
+
+    def fetchone(query: str, parameters: tuple) -> tuple:
+        return cursor.execute(query, parameters).fetchone()
+
+    def fetchall(query: str, parameters: tuple) -> list[tuple]:
+        return cursor.execute(query, parameters).fetchall()
+
+    return {"execute": execute, "fetchone": fetchone, "fetchall": fetchall}
